requirements.txt
customtkinter
pypresence
Pillow
# pygame
# requests
# gitpython